add multiple dosiser and search patients

+++ application/models/Dossier_model.php
@@ -22,7 +22,7 @@ class Dossier_model extends CI_Model
         SELECT d.*, 
                CONCAT(p.first_name, ' ', p.last_name) AS full_name, 
                 GROUP_CONCAT(a.analyse_name SEPARATOR ', ') AS analyses_names,
-               p.age, p.phone, p.description, 
+               p.age, p.phone, 
                SUM(a.analyse_price) as price,
                DATE_FORMAT(d.date_update, '%d/%m/%Y') AS update_date
         FROM dossiers as d
@@ -61,7 +61,7 @@ class Dossier_model extends CI_Model
         SELECT d.*, 
                CONCAT(p.first_name, ' ', p.last_name) AS full_name, 
                 GROUP_CONCAT(a.analyse_name SEPARATOR ', ') AS analyses_names,
-               p.age, p.phone, p.description,l.labo_name, 
+               p.age, p.phone,l.labo_name, 
                SUM(a.analyse_price) as price,
                DATE_FORMAT(d.date_update, '%d/%m/%Y') AS update_date
         FROM dossiers as d
@@ -80,7 +80,7 @@ class Dossier_model extends CI_Model
 
         // Group by, order by, and limit
         $queryStr .= "
-        GROUP BY d.patient_id
+        GROUP BY d.ref
         ORDER BY d.id DESC
         LIMIT ?, ?";
 
@@ -101,25 +101,36 @@ class Dossier_model extends CI_Model
      * This function is used to add new dossier to system
      * @return number $insert_id : This is last inserted id
      */
-    function addNewDossier($analyses,$patientInfo,$labo,$dossiertInfo)
+    function addNewDossier($analyses,$patientInfo,$labo,$dossiertInfo,$patientId)
     {
         $this->db->trans_start();
-        $this->db->insert('patients', $patientInfo);
 
-        $patient_id = $this->db->insert_id();
+
+        $query = $this->db->get_where('patients', ['id' => $patientId]);
+
+        if ($query->num_rows() > 0) {
+            // Update existing record
+            $this->db->where('id', $patientId);
+            $this->db->update('patients', $patientInfo);
+            $patient_id = $patientId;
+        } else {
+            $this->db->insert('patients', $patientInfo);
+            $patient_id = $this->db->insert_id();
+        }
 
         $_date = DateTime::createFromFormat('d-m-Y', $dossiertInfo->date)->format('Y-m-d');
-
+        $ref = substr(md5(uniqid(mt_rand(), true)), 0, 10);
         foreach ($analyses as $analyse){
             $data = [
+                'ref'=>$ref,
                 'patient_id'=>$patient_id,
                 'analyse_id'=>$analyse,
                 'labo_id'=>$labo,
                 'date_update'=>$_date,
                 'reduction'=>$dossiertInfo->reduction,
+                'description'=>$dossiertInfo->description,
             ];
             $this->db->insert('dossiers', $data);
-
         }
 
         $insert_id = $this->db->insert_id();
@@ -135,12 +146,12 @@ class Dossier_model extends CI_Model
      * @return array $result : This is dossier information
      */
 
-    function getDossierInfo($patient_id)
+    function getDossierInfo($ref)
     {
-        $this->db->select('a.*, p.first_name, p.last_name,p.age, p.phone, p.description',);
-        $this->db->from('dossiers as a');
-        $this->db->join('patients as p', 'a.patient_id = p.id',);
-        $this->db->where('a.patient_id', $patient_id);
+        $this->db->select('d.*, p.first_name, p.last_name,p.age, p.phone',);
+        $this->db->from('dossiers as d');
+        $this->db->join('patients as p', 'd.patient_id = p.id',);
+        $this->db->where('d.ref', $ref);
         $this->db->limit(1);
         $query = $this->db->get();
         $data = $query->row();
@@ -148,10 +159,10 @@ class Dossier_model extends CI_Model
         return $data;
     }
 
-    public function getSelectedAnalyses($patient_id) {
+    public function getSelectedAnalyses($ref) {
         $this->db->select('analyse_id');
         $this->db->from('dossiers');
-        $this->db->where('patient_id', $patient_id);
+        $this->db->where('ref', $ref);
         $query = $this->db->get();
         return array_column($query->result_array(), 'analyse_id');
     }
@@ -163,7 +174,7 @@ class Dossier_model extends CI_Model
      * @param array $dossierInfo : This is dossier updated information
      * @param number $dossierId : This is dossier id
      */
-    function editDossier($analyses, $patientInfo,$patientId,$labo,$dossiertInfo)
+    function editDossier($analyses, $patientInfo,$patientId,$labo,$dossiertInfo,$ref)
     {
         // Start the transaction
         $this->db->trans_start();
@@ -175,7 +186,7 @@ class Dossier_model extends CI_Model
 
 
         // Delete existing dossiers for this patient.
-        $this->db->where('patient_id', $patientId);
+        $this->db->where('ref', $ref);
         $this->db->delete('dossiers');
 
         if ($this->db->affected_rows() === 0) {
@@ -183,14 +194,17 @@ class Dossier_model extends CI_Model
             return FALSE;
         }
         $_date = DateTime::createFromFormat('d-m-Y', $dossiertInfo->date)->format('Y-m-d');
+        $ref = substr(md5(uniqid(mt_rand(), true)), 0, 10);
         if (!empty($analyses)) {
             foreach ($analyses as $analyse) {
                 $data = [
+                    'ref' => $ref,
                     'patient_id' => $patientId,
                     'analyse_id' => $analyse,
                     'labo_id' => $labo,
                     'date_update' => $_date,
                     'reduction' => $dossiertInfo->reduction,
+                    'description' => $dossiertInfo->description,
                 ];
                 $this->db->insert('dossiers', $data);
             }
@@ -201,12 +215,17 @@ class Dossier_model extends CI_Model
         return $this->db->trans_status();
     }
 
-    function deleteDossier($patientId)
+    function deleteDossier($ref,$patientId)
     {
-        $this->db->where('patient_id', $patientId);
+        $this->db->where('ref', $ref);
         $this->db->delete('dossiers');
-        $this->db->where('id', $patientId);
-        $this->db->delete('patients');
+
+        $query = $this->db->get_where('dossiers', ['patient_id' => $patientId]);
+
+        if($query->num_rows() == 0){
+            $this->db->where('id', $patientId);
+            $this->db->delete('patients');
+        }
 
         return $this->db->affected_rows();
     }
@@ -236,7 +255,7 @@ class Dossier_model extends CI_Model
         SELECT d.*, 
                CONCAT(p.first_name, ' ', p.last_name) AS full_name, 
                GROUP_CONCAT(a.analyse_name SEPARATOR ', ') AS analyses_names,
-               p.age, p.phone, p.description, 
+               p.age, p.phone, 
                SUM(a.analyse_price) - d.reduction AS price,
                DATE_FORMAT(d.date_update, '%d/%m/%Y') AS update_date
         FROM dossiers AS d
@@ -244,7 +263,7 @@ class Dossier_model extends CI_Model
         JOIN patients AS p ON p.id = d.patient_id
         JOIN laboratoires AS l ON l.id = d.labo_id
         WHERE d.date_update BETWEEN ? AND ? and d.labo_id=$labo_id
-        GROUP BY d.patient_id";
+        GROUP BY d.ref ORDER BY d.id DESC";
 
         $query = $this->db->query($queryStr, [$startDateTime, $endDateTime]);
 
@@ -257,6 +276,13 @@ class Dossier_model extends CI_Model
         $this->db->from('laboratoires');
         $query = $this->db->get();
         return $query->result();
+    }
+
+    public function searchPatients($query) {
+        $this->db->like('first_name', $query);
+        $this->db->or_like('last_name', $query);
+        $this->db->limit(5);
+        return $this->db->get('patients')->result();
     }
 
 }
